requirements.txt
argparse
datasets
numpy
openai
peft
python-dotenv
torch
tqdm
transformers
trl
wandb